# openapi/avedra-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AVEDRA API",
    "description": "API REST de AVEDRA — Asistencia, Convivencia y Evaluacion",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/evaluacion/planilla": {
      "get": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Obtener Planilla",
        "description": "Retorna la planilla de notas del grupo con definitivas calculadas.",
        "operationId": "obtener_planilla_api_v1_evaluacion_planilla_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "asignacion_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asignacion Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanillaEstudianteResponse"
                  },
                  "title": "Response Obtener Planilla Api V1 Evaluacion Planilla Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluacion/notas-masivas": {
      "post": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Registrar Notas Masivas",
        "description": "Registra notas para multiples estudiantes en una actividad.",
        "operationId": "registrar_notas_masivas_api_v1_evaluacion_notas_masivas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrarNotasMasivasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/evaluacion/siee": {
      "get": {
        "tags": [
          "evaluacion"
        ],
        "summary": "Get Configuracion Siee",
        "description": "Retorna la configuracion SIEE del anio lectivo.",
        "operationId": "get_configuracion_siee_api_v1_evaluacion_siee_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "anio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Anio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionSIEEResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/informes/boletin-periodo/{estudiante_id}": {
      "get": {
        "tags": [
          "informes"
        ],
        "summary": "Boletin Periodo",
        "description": "Genera el boletin de un estudiante para un periodo (PDF o Excel).",
        "operationId": "boletin_periodo_api_v1_informes_boletin_periodo__estudiante_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estudiante_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Estudiante Id"
            }
          },
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pdf",
              "title": "Formato"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/informes/boletin-anual/{estudiante_id}": {
      "get": {
        "tags": [
          "informes"
        ],
        "summary": "Boletin Anual",
        "description": "Genera el boletin anual de un estudiante (PDF o Excel).",
        "operationId": "boletin_anual_api_v1_informes_boletin_anual__estudiante_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estudiante_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Estudiante Id"
            }
          },
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "anio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Anio Id"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pdf",
              "title": "Formato"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/informes/notas": {
      "get": {
        "tags": [
          "informes"
        ],
        "summary": "Informe Notas",
        "description": "Genera el informe de notas del grupo (Excel o PDF).",
        "operationId": "informe_notas_api_v1_informes_notas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "asignacion_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asignacion Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          },
          {
            "name": "fecha_desde",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Desde"
            }
          },
          {
            "name": "fecha_hasta",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Hasta"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "excel",
              "title": "Formato"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/informes/asistencia": {
      "get": {
        "tags": [
          "informes"
        ],
        "summary": "Informe Asistencia",
        "description": "Genera el informe de asistencia del grupo (Excel o PDF).",
        "operationId": "informe_asistencia_api_v1_informes_asistencia_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "asignacion_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asignacion Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          },
          {
            "name": "fecha_desde",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Desde"
            }
          },
          {
            "name": "fecha_hasta",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Hasta"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "excel",
              "title": "Formato"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfiguracionSIEEResponse": {
        "properties": {
          "anio_id": {
            "type": "integer",
            "title": "Anio Id"
          },
          "modo": {
            "type": "string",
            "title": "Modo",
            "default": "LIBRE"
          }
        },
        "type": "object",
        "required": [
          "anio_id"
        ],
        "title": "ConfiguracionSIEEResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanillaEstudianteResponse": {
        "properties": {
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "nombre_completo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Completo"
          },
          "definitiva": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definitiva"
          },
          "promedio_ajustado": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promedio Ajustado"
          },
          "notas": {
            "title": "Notas"
          }
        },
        "type": "object",
        "required": [
          "estudiante_id"
        ],
        "title": "PlanillaEstudianteResponse"
      },
      "RegistrarNotaItem": {
        "properties": {
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "actividad_id": {
            "type": "integer",
            "title": "Actividad Id"
          },
          "valor": {
            "type": "number",
            "title": "Valor"
          }
        },
        "type": "object",
        "required": [
          "estudiante_id",
          "actividad_id",
          "valor"
        ],
        "title": "RegistrarNotaItem"
      },
      "RegistrarNotasMasivasRequest": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "asignacion_id": {
            "type": "integer",
            "title": "Asignacion Id"
          },
          "periodo_id": {
            "type": "integer",
            "title": "Periodo Id"
          },
          "anio_id": {
            "type": "integer",
            "title": "Anio Id"
          },
          "actividad_id": {
            "type": "integer",
            "title": "Actividad Id"
          },
          "notas": {
            "items": {
              "$ref": "#/components/schemas/RegistrarNotaItem"
            },
            "type": "array",
            "title": "Notas"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "asignacion_id",
          "periodo_id",
          "anio_id",
          "actividad_id",
          "notas"
        ],
        "title": "RegistrarNotasMasivasRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}